+++ changelog.md
@@ -8,3 +8,8 @@
 
 ## 0.4.1
 - Add support for intersection tests with arbitrary polygons. [Issue #49](https://github.com/myelin-ai/engine/issues/49)
+
+## 0.5.0
+- Move `own_description` from the signature of `ObjectBehavior::step` to the new function `WorldInteractor::own_object(&self) -> Object<'_>`
+    - This also allows access to an object's own ID
+- Support object retrieval with an ID via `Simulation::object(&self, id: Id) -> Option<Object<'_>>`